// backend/src/modules/auth/enums/auth.status-code.enum.ts
export enum EnumAuthStatusCodeError {
  jwtAccessTokenInvalid = 5120,
  jwtRefreshTokenInvalid = 5121,
  socialGoogleRequired = 5122,
  socialGoogleInvalid = 5123,
  socialAppleRequired = 5124,
  socialAppleInvalid = 5125,
  twoFactorInvalid = 5126,
  twoFactorChallengeInvalid = 5127,
  twoFactorNotEnabled = 5128,
  twoFactorAlreadyEnabled = 5129,
  twoFactorRequiredSetup = 5132,
  twoFactorNotRequiredSetup = 5133,
  twoFactorAttemptTemporaryLock = 5134,
}