// backend/src/modules/user/enums/user.status-code.enum.ts
export enum EnumUserStatusCodeError {
  notFound = 5150,
  notSelf = 5151,
  emailExist = 5152,
  usernameExist = 5153,
  mobileNumberNotFound = 5154,
  statusInvalid = 5155,
  blockedInvalid = 5156,
  inactiveForbidden = 5157,
  deletedForbidden = 5158,
  blockedForbidden = 5159,
  passwordNotMatch = 5160,
  passwordMustNew = 5161,
  passwordExpired = 5162,
  passwordAttemptMax = 5163,
  mobileNumberInvalid = 5164,
  usernameNotAllowed = 5165,
  usernameContainBadWord = 5166,
  emailNotVerified = 5167,
  loginWithNotSupported = 5168,
  passwordNotSet = 5169,
  tokenInvalid = 5170,
  emailAlreadyVerified = 5171,
  mobileNumberExist = 5172,
  verificationEmailResendLimitExceeded = 5173,
  forgotPasswordRequestLimitExceeded = 5174,
}